# small encoding allocation optimisation