Profiling with dtrace showed that the current code is doing quite a lot of reallocations, coming from the `from_iter` call.

This PR brings a ~10-20% performance improvement for the encoding part, in terms of total duration.

Benchmarked this change with kagome's benchmark.

`num_validators = 100`

Previously, on my M2 Pro with 32 Gib RAM, results were: 
```
~~~ [ Benchmark case: 303 bytes ] ~~~
Encode RUST (100 cycles): 846 us
Decode RUST (100 cycles): 55.689 ms
Encode C++ (100 cycles): 519 us
Decode C++ (100 cycles): 30.949 ms

~~~ [ Benchmark case: 5007 bytes ] ~~~
Encode RUST (100 cycles): 11.092 ms
Decode RUST (100 cycles): 73.303 ms
Encode C++ (100 cycles): 5.334 ms
Decode C++ (100 cycles): 42.614 ms

~~~ [ Benchmark case: 100015 bytes ] ~~~
Encode RUST (100 cycles): 215.752 ms
Decode RUST (100 cycles): 428.894 ms
Encode C++ (100 cycles): 100.325 ms
Decode C++ (100 cycles): 280.875 ms

~~~ [ Benchmark case: 1000015 bytes ] ~~~
Encode RUST (100 cycles): 2154.09 ms
Decode RUST (100 cycles): 3880.8 ms
Encode C++ (100 cycles): 1001.65 ms
Decode C++ (100 cycles): 2552.96 ms

~~~ [ Benchmark case: 10000015 bytes ] ~~~
Encode RUST (100 cycles): 29.2552 s
Decode RUST (100 cycles): 39.0674 s
Encode C++ (100 cycles): 15.8985 s
Decode C++ (100 cycles): 26.2342 s
```

With this PR, results are:
```
~~~ [ Benchmark case: 303 bytes ] ~~~
Encode RUST (100 cycles): 748 us
Decode RUST (100 cycles): 56.005 ms
Encode C++ (100 cycles): 530 us
Decode C++ (100 cycles): 30.887 ms

~~~ [ Benchmark case: 5007 bytes ] ~~~
Encode RUST (100 cycles): 8.986 ms
Decode RUST (100 cycles): 73.672 ms
Encode C++ (100 cycles): 5.406 ms
Decode C++ (100 cycles): 42.62 ms

~~~ [ Benchmark case: 100015 bytes ] ~~~
Encode RUST (100 cycles): 172.488 ms
Decode RUST (100 cycles): 429.697 ms
Encode C++ (100 cycles): 99.037 ms
Decode C++ (100 cycles): 279.038 ms

~~~ [ Benchmark case: 1000015 bytes ] ~~~
Encode RUST (100 cycles): 1723.59 ms
Decode RUST (100 cycles): 3842.43 ms
Encode C++ (100 cycles): 1008.16 ms
Decode C++ (100 cycles): 2571.68 ms

~~~ [ Benchmark case: 10000015 bytes ] ~~~
Encode RUST (100 cycles): 25.0707 s
Decode RUST (100 cycles): 39.4162 s
Encode C++ (100 cycles): 16.0018 s
Decode C++ (100 cycles): 26.1704 s
```